requirements.txt
streamlit
pandas
fpdf2
plotly
Pillow